requirements.txt
# torch==2.5.0
# torchvision==0.20.0
pyyaml==6.0.2
facenet-pytorch==2.6.0
opencv-python==4.10.0.84
numpy==1.26.0
insightface==0.7.3
albumentations==1.4.18
tf_keras==2.17.0
onnxruntime==1.19.2
colorama==0.4.6
